requirements.txt
albumentations==1.3.1
numpy>=1.24.2,<2.0
onnxruntime==1.16.3
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
Pillow==10.1.0
matplotlib==3.10.6
